requirements.txt
pandas
Flask
Werkzeug
rembg
numpy
tensorflow
keras-cv
Pillow
scikit-image
joblib
scikit-learn